+++ packages/commercetools/theme/middleware.config.js
@@ -1,3 +1,5 @@
+const { megaMenuCategoriesQuery } = require('./queries');
+
 module.exports = {
   integrations: {
     ct: {
@@ -26,34 +28,7 @@ module.exports = {
         country: 'US'
       },
       customQueries: {
-        'megamenu-categories-query': ({ variables }) => {
-
-          const megaMenuCategoriesQuery = `
-            fragment DefaultRootCategory on Category {
-              id
-              slug(acceptLanguage: $acceptLanguage)
-              name(acceptLanguage: $acceptLanguage)
-              childCount
-            }
-            query categories($where: String, $sort: [String!], $limit: Int, $offset: Int, $acceptLanguage: [Locale!]) {
-              categories(where: $where, sort: $sort, limit: $limit, offset: $offset) {
-                offset
-                count
-                total
-                results {
-                  ...DefaultRootCategory
-                }
-              }
-            }
-          `;
-
-          variables.where = 'parent is not defined';
-
-          return {
-            query: megaMenuCategoriesQuery,
-            variables
-          };
-        }
+        'megamenu-categories-query': megaMenuCategoriesQuery
       }
     }
   }
